+++ app/navigation/AppNavigator.js
@@ -11,9 +11,18 @@ export default function App() {
   return (
     <NavigationContainer>
       <Drawer.Navigator initialRouteName="Home">
-        <Drawer.Screen name="Home" component={HomeScreen} />
-        <Drawer.Screen name="Creature" component={CreatureScreen} />
-        <Drawer.Screen name="Spell" component={SpellScreen} />
+        <Drawer.Screen
+            name="Home"
+            component={HomeScreen}
+        />
+        <Drawer.Screen
+            name="Creature"
+            component={CreatureScreen}
+        />
+        <Drawer.Screen
+            name="Spell"
+            component={SpellScreen}
+        />
       </Drawer.Navigator>
     </NavigationContainer>
   );