+++ app/screens/Creature/Styles.js
@@ -2,17 +2,35 @@ import { StyleSheet } from 'react-native';
 
 const Styles = StyleSheet.create({
     container: {
-        flex: 1,
-        alignItems: 'center',
-        marginTop: 20,
+        paddingTop: 50,
     },
     title: {
+        textAlign: 'center',
         fontSize: 23,
-        marginBottom: 20,
     },
     description: {
+        textAlign: 'center',
         fontSize: 18,
         marginTop: 20,
+        marginBottom: 10,
+    },
+    card: {
+        backgroundColor: "rgba(0, 0, 0, 0.1)",
+        borderRadius: 10,
+        padding: 10,
+        margin: 10,
+    },
+    image: {
+        borderRadius: 10,
+        padding: 10,
+        margin: 10,
+        alignItems: 'center',
+    },
+    info: {
+        // backgroundColor: "rgba(0, 0, 0, 0.1)",
+        borderRadius: 10,
+        padding: 10,
+        margin: 10,
     }
 });