+++ app/screens/Home/Index.js
@@ -20,4 +20,5 @@ const HomeScreen = () => {
     </View>
   );
 };
+
 export default HomeScreen;