+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Change the home screen.
 - Add the creature screen.
 - Add the spell screen and return data from API.
+- Add the creature screen and return data from API.
 
 #v0.1.0